server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.6",
    "loguru>=0.7.3",
    "openai>=1.58.1",
    "pgvector>=0.3.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
